Move benchmark execution method into benchmark macro
